Added some functionality to one of the graphs
I tested a method of getting some sort of data one of the graphs. It
generates a random value between 1 and 6 and plots it on the graph every
second. Must try to get the graph to move along the x axis to the left
and must find a way to store data points.