I GOT THE GRAPH TO WORK
So I got the graphs to work properly. The solution is only implemented
on the temperature chart, but can be implemented in the others. There is
also a "Start Simulation" button in the file menu, but there is no code
in its method. If we decide it, we can have the simulation start on
button click or when the program starts up. That would be a group
decision but it can be done.