Graphs are set up to work
All the Graphs are setup to work properly now. Now what must be fine
tuned is the generator classes, in order to get accurate and realistic
values. The Blood Pressure chart has been excluded because there has to
be a discussion about the units of measure and plotting for blood
pressure. However the blood pressure chart works like all the rest.